removed restRequests folder and changed file extensions in srvice folder to be only ts
