+++ src/components/campaigns/InlineDonation.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'next-i18next'
 import { CampaignResponse } from 'gql/campaigns'
 import { baseUrl, routes } from 'common/routes'
 import { money } from 'common/util/money'
-import { useSinglePriceList, useDonationSession } from 'common/hooks/donations'
+import { useSinglePriceList, useDonationSession } from 'common/hooks/donation'
 import LinkButton from 'components/common/LinkButton'
 import CampaignProgress from './CampaignProgress'
 import DonorsAndDonations from './DonorsAndDonations'
