+++ src/components/cities/CreateForm.tsx
@@ -20,7 +20,7 @@ import { CityFormData, CityInput, CityResponse } from 'gql/cities'
 import { useCountriesList } from 'common/hooks/cities'
 import { routes } from 'common/routes'
 import { ApiErrors } from 'service/apiErrors'
-import { useCreateCity } from 'service/restRequests/city'
+import { useCreateCity } from 'service/city'
 import { AlertStore } from 'stores/AlertStore'
 import GenericForm from 'components/common/form/GenericForm'
 import FormTextField from 'components/common/form/FormTextField'
