+++ src/components/cities/modals/DeleteModal.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'next-i18next'
 import { CityResponse } from 'gql/cities'
 import { ApiErrors } from 'service/apiErrors'
 import { endpoints } from 'service/apiEndpoints'
-import { useDeleteCity } from 'service/restRequests/city'
+import { useDeleteCity } from 'service/city'
 import { ModalStore } from 'stores/documents/ModalStore'
 import { AlertStore } from 'stores/AlertStore'
 import { useRouter } from 'next/router'
