+++ src/components/countries/CountryForm.tsx
@@ -11,7 +11,7 @@ import { makeStyles } from '@mui/styles'
 import { routes } from 'common/routes'
 import { AlertStore } from 'stores/AlertStore'
 import { ApiErrors, isAxiosError, matchValidator } from 'service/apiErrors'
-import { useCreateCountry, useEditCountry } from 'service/restRequests/country'
+import { useCreateCountry, useEditCountry } from 'service/country'
 import { CountryInput, CountryResponse } from 'gql/countries'
 import { name } from 'common/form/validation'
 import GenericForm from 'components/common/form/GenericForm'
