+++ src/components/documents/Form.tsx
@@ -11,7 +11,7 @@ import { DocumentInput, DocumentResponse } from 'gql/document'
 import { useDocument } from 'common/hooks/documents'
 import { routes } from 'common/routes'
 import { ApiErrors } from 'service/apiErrors'
-import { useCreateDocument, useEditDocument } from 'service/restRequests/document'
+import { useCreateDocument, useEditDocument } from 'service/document'
 import { AlertStore } from 'stores/AlertStore'
 import GenericForm from 'components/common/form/GenericForm'
 import FormTextField from 'components/common/form/FormTextField'
