+++ src/components/documents/grid/DeleteAllModal.tsx
@@ -9,7 +9,7 @@ import { useTranslation } from 'next-i18next'
 import { DocumentResponse } from 'gql/document'
 import { ApiErrors } from 'service/apiErrors'
 import { endpoints } from 'service/apiEndpoints'
-import { useDeleteManyDocuments } from 'service/restRequests/document'
+import { useDeleteManyDocuments } from 'service/document'
 import { ModalStore } from 'stores/documents/ModalStore'
 import { AlertStore } from 'stores/AlertStore'
