+++ src/components/vaults/Form.tsx
@@ -11,7 +11,7 @@ import { VaultInput, VaultResponse } from 'gql/vault'
 import { useVault } from 'common/hooks/vaults'
 import { routes } from 'common/routes'
 import { ApiErrors } from 'service/apiErrors'
-import { useCreateVault, useEditVault } from 'service/restRequests/vault'
+import { useCreateVault, useEditVault } from 'service/vault'
 import { AlertStore } from 'stores/AlertStore'
 import GenericForm from 'components/common/form/GenericForm'
 import FormTextField from 'components/common/form/FormTextField'
