--- src/service/restRequests/contact.tsx
@@ -1,12 +1,0 @@
-import { AxiosResponse } from 'axios'
-
-import { apiClient } from 'service/apiClient'
-import { endpoints } from 'service/apiEndpoints'
-import { ContactRequestResponse, ContactRequestInput } from 'gql/contact'
-
-export const createContactRequest = async (data: ContactRequestInput) => {
-  return await apiClient.post<ContactRequestInput, AxiosResponse<ContactRequestResponse>>(
-    endpoints.support.createInfoRequest.url,
-    data,
-  )
-}
